+++ src/engine/core/examples/print_default_config.rs
@@ -1,0 +1,6 @@
+fn main() {
+    println!(
+        "{}",
+        serde_yaml::to_string(&kime_engine_core::RawConfig::default()).unwrap()
+    );
+}